Initial svelte version
- draggy annotations

#5455
